Hangar Power Outage Fix (#390)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

I watched as two players incorrectly added onto their ship while it was
inside a hangar... The result when they left was that the entire
hangar... All of them... Was now dependent on this SINGLE APC for
power... Well without a power source all hangars lost power
immediately...

This is an extension of the problem involving ships with area's that
lack APCs but on a much larger grand scale...


Now any attempt to place an APC inside the hangar will result in a "this
area already has an APC"


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Prevents people from breaking hangars.... in this one way.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Centcom now bluespaces power to hangars
fix: Stops players from being able to place APCs in hangars
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->