Mothroach Mission Reflavor (#402)
About The Pull Request
Changes the name and description of the "recover mothroach" mission.

Why It's Good For The Game
Honestly, the previous version is just bad. And while at the time I was
alright with it... People mentioned it and I agree with what they felt
so its being reflavored completely.

Changelog
🆑
fix: Reflavors the mothroach recovery mission
/:cl: