Edits some Space Ruin spawn weights (#396)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This reduces the cost of the Pentest Space Ruins to match that of
Shiptests

Also increases the ruin cost for the Scrap Station Ruin as it seems to
be appearing all the time.


## Why It's Good For The Game

Ruin variety and less people dying.... Part of the issue could be that a
mission is tied to the scrap station and because of people taking a
mission and never finishing it keeps trying to give it to players... not
sure yet.

## Changelog

:cl:
balance: Makes some space ruins more common and others less....
hopefully.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->