Ship: Proteus-class Decommissioned Frigate (#310)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds the Proteus-Class Decommissioned Frigate, a CLIP fast response ship
that has been stripped down and sold off into private hands. The captain
is either a ship restorationist, some kind of military larper, a retired
CLIP officer, or maybe just a merc looking for a fast, milspec vessel to
take them to their next paycheck. Why the captain owns the ship is left
ambiguous to encourage roleplay. The crew is mostly civilians and mercs,
presumably hired on to either help restore the vessel and/or use it to
make money.

The concept was largely derived from the Arke, which mentions the
Proteus. Hence the Arke is the scaled-down, specialized variant, while
the Proteus is the larger, generalist vessel.

Notables:
Much of the ship has been stripped down, probably in a hurry, but
certain things were overlooked.
The vessel does not bear any CLIP markings, or any markings, really.
Lots of bare console/machine frames to upgrade later.
Starboard engine has room for expansion via a utility area.
Two high-performance ion thrusters with two regulars, making the ship
very fast.
Functional atmos system with a mixer/vent (I've tested this, but I'm not
super confident in my atmos expertise).
Functional camera system with coverage over the entire ship.
Four functional turrets, linked to a control set on the bridge.
Hand drill and industrial welder for fluff.
Surgical bag in the medical locker as I think it fits the narrative of
rapid response.
A bit of miner gear, including a mining hardsuit, probably there in an
attempt to re-purpose the vessel.
Mecha charging station with a fabricator board in a nearby trash can. No
actual powerstation board, gotta get one.
Surplus CM-5 and a brimstone in the armory, along with a box of ammo for
each. Also a CM23 pistol in the captain's quarters.
Set of armor and a CLIP vest, also in the armory.
Jetpack and a basic space suit near the external airlock.

Special:
CLIP telecomms relay preset + blackbox, left forgotten on the bridge.
Captain's locker on the bridge contains CLIP officer stuff in case the
captain wants to roleplay in that direction.
Hidden safe with a random amount of up to 1500 cash and a Paladin
conversion kit.
Super cool punching bag.

**Hopefully I've configured the static ports correctly, same with the
atmos system. Needs testing.**

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Screenshots
![Proteus
1](https://github.com/user-attachments/assets/c752a7f5-6b45-4dbb-b232-f28177de4a0d)
![Proteus
2](https://github.com/user-attachments/assets/2c586e1b-2e63-4d18-afd8-71dc5a03a4fd)

## Why It's Good For The Game

I think it's pretty cool idea that hasn't been done yet. This ship in
particular is good for a duo who wants to do missions, with plenty of
space to install things like an RnD setup, a field hospital, or some
other thing. It might get crowded with three or more people though,
which is sort of on purpose. The Proteus-class is very confining, packed
with military efficiency: from the small but very robust cargo area to
the compact crew quarters. It's designed to be a little suffocating.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added the Proteus-class, a decommissioned CLIP frigate, now in
private hands.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Ossa88 (SYNAPSE) <MegaMan101@gmail.com>