Entertainment Telescreen Fix (#463)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Due to an oversight, Entertainment Telescreens were given a
'interaction_flags_atom = NONE' which removed all interaction that a
Carbon character could do to the monitor.

This essentially made the monitor only an announcement board should
another crew somewhere use a broadcast camera.

YET, at the same time, the Entertainment Telescreen was given a "big
click" bypass, which was SUPPOSED to allow players to interact with the
monitor from +1 more distance away (as technically the monitor is in the
square next to the wall its slapped onto).

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

What was the purpose of the Entertainment Telescreen if not to watch a
broadcast.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: Entertainment Telescreens are actually interactable by Carbons and
are no longer exclusive to Robots and AIs.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->