Science Silicons can now Pick Up Mecha Ammo (#484)
Fixes an issue in which Science Silicons could not pick up and apply
mecha/exosuit ammunition boxes with the science apparatus.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR adds /obj/item/mecha_ammo to the storable list that the science
silicon's apparatus can pick up.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Science silicons are able to construct and maintain mechs/exosuits.
However, they lack the capacity to resupply the ballistic weaponry that
can be mounted onto the exosuits. This limits silicon-centered ships,
and would likely result in the crew preferring laser weapons over
ballistics. This PR rectifies this, allowing these types of crews more
freedom in their robotics arsenal, and provides just a bit more autonomy
to silicons.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: Resolves an oversight that prevented science silicons from picking
up mecha ammo boxes.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->